+++ apps/meteor/client/views/room/contextualBar/Info/EditRoomInfo/useEditRoomInitialValues.ts
@@ -1,28 +1,12 @@
 import type { IRoomWithRetentionPolicy } from '@rocket.chat/core-typings';
-import { useSetting } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 
 import { roomCoordinator } from '../../../../../lib/rooms/roomCoordinator';
-
-const getPolicyRoomType = (roomType: IRoomWithRetentionPolicy['t']) => {
-	switch (roomType) {
-		case 'c':
-			return 'Channels';
-		case 'p':
-			return 'Groups';
-		case 'd':
-			return 'DMs';
-	}
-};
+import { useRetentionPolicy } from '../../../hooks/useRetentionPolicy';
 
 export const useEditRoomInitialValues = (room: IRoomWithRetentionPolicy) => {
+	const retentionPolicy = useRetentionPolicy(room);
 	const { t, ro, archived, topic, description, announcement, joinCodeRequired, sysMes, encrypted, retention, reactWhenReadOnly } = room;
-
-	const retentionPolicyEnabled = useSetting('RetentionPolicy_Enabled');
-	const maxAgeDefault = useSetting<number>(`RetentionPolicy_MaxAge_${getPolicyRoomType(room.t)}`) || 30;
-	const retentionEnabledDefault = useSetting<boolean>(`RetentionPolicy_AppliesTo${getPolicyRoomType(room.t)}`);
-	const excludePinnedDefault = useSetting('RetentionPolicy_DoNotPrunePinned');
-	const filesOnlyDefault = useSetting('RetentionPolicy_FilesOnly');
 
 	return useMemo(
 		() => ({
@@ -40,25 +24,21 @@ export const useEditRoomInitialValues = (room: IRoomWithRetentionPolicy) => {
 			systemMessages: Array.isArray(sysMes) ? sysMes : [],
 			hideSysMes: Array.isArray(sysMes) ? !!sysMes?.length : !!sysMes,
 			encrypted,
-			...(retentionPolicyEnabled && {
-				retentionEnabled: retention?.enabled ?? retentionEnabledDefault,
+			...(retentionPolicy && {
+				retentionEnabled: retention?.enabled ?? Boolean(retentionPolicy),
 				retentionOverrideGlobal: !!retention?.overrideGlobal,
-				retentionMaxAge: Math.min(retention?.maxAge, maxAgeDefault) || maxAgeDefault,
-				retentionExcludePinned: retention?.excludePinned ?? excludePinnedDefault,
-				retentionFilesOnly: retention?.filesOnly ?? filesOnlyDefault,
+				retentionMaxAge: retention?.maxAge ?? retentionPolicy.maxAge,
+				retentionExcludePinned: retention?.excludePinned ?? retentionPolicy.excludePinned,
+				retentionFilesOnly: retention?.filesOnly ?? retentionPolicy.filesOnly,
 			}),
 		}),
 		[
 			announcement,
 			archived,
 			description,
-			excludePinnedDefault,
-			filesOnlyDefault,
 			joinCodeRequired,
-			maxAgeDefault,
 			retention,
-			retentionEnabledDefault,
-			retentionPolicyEnabled,
+			retentionPolicy,
 			ro,
 			room,
 			sysMes,
